test/integration: Shuffle code into more sensible directories
