+++ test/integration/gles/replay/reference/one-depth.png
