+++ test/integration/gles/replay/reference/solid-black.png
