+++ test/integration/gles/replay/reference/solid-blue.png
