+++ test/integration/gles/replay/reference/solid-green.png
