+++ test/integration/gles/replay/reference/solid-red.png
