+++ test/integration/gles/replay/reference/textured-square.png
