+++ test/integration/gles/replay/reference/triangle-180.png
