+++ test/integration/gles/replay/reference/triangle-depth.png
