+++ test/integration/gles/replay/reference/triangle.png
