+++ test/integration/gles/replay/reference/triangle_2.png
