+++ test/integration/gles/replay/reference/undef-fb.png
