+++ test/integration/gles/snippets/clear_backbuffer.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package samples
+package snippets
 
 import (
 	"context"
